[CHORE] feed 페이지로 분리 후 연결된 링크 수정 #32
